+++ include/libasyik/http.hpp
@@ -1,6 +1,9 @@
 #ifndef LIBASYIK_ASYIK_HTTP_HPP
 #define LIBASYIK_ASYIK_HTTP_HPP
 
+// Main HTTP header - includes all modular headers for backward compatibility
+// This file now serves as a facade that includes all split headers
+
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/any.hpp>
 #include <boost/asio/ssl/error.hpp>
@@ -23,6 +26,12 @@
 #include "internal/digestauth.hpp"
 #include "service.hpp"
 
+// Include modular headers
+#include "http_client.hpp"
+#include "http_server.hpp"
+#include "http_types.hpp"
+#include "http_websocket.hpp"
+
 namespace fibers = boost::fibers;
 using fiber = boost::fibers::fiber;
 using tcp = boost::asio::ip::tcp;
@@ -31,548 +40,12 @@ namespace ssl = asio::ssl;
 namespace beast = boost::beast;
 namespace ip = boost::asio::ip;
 namespace websocket = beast::websocket;
+namespace http = boost::beast::http;
 
 namespace asyik {
 void _TEST_invoke_http();
 
-static const size_t default_request_body_limit = 1 * 1024 * 1024;
-static const size_t default_request_header_limit = 1 * 1024 * 1024;
-static const size_t default_response_body_limit = 64 * 1024 * 1024;
-static const size_t default_response_header_limit = 1 * 1024 * 1024;
-
-using http_route_args = std::vector<std::string>;
-
-using http_beast_request =
-    boost::beast::http::request<boost::beast::http::string_body>;
-using http_beast_response =
-    boost::beast::http::response<boost::beast::http::string_body>;
-using http_request_headers = http_beast_request::header_type;
-using http_request_body = http_beast_request::body_type::value_type;
-using http_response_headers = http_beast_response::header_type;
-using http_response_body = http_beast_response::body_type::value_type;
-
-using http_route_callback =
-    std::function<void(http_request_ptr, const http_route_args&)>;
-using http_route_tuple =
-    std::tuple<std::string, std::regex, http_route_callback>;
-
-using websocket_route_callback =
-    std::function<void(websocket_ptr, const http_route_args&)>;
-using websocket_route_tuple =
-    std::tuple<std::string, std::regex, websocket_route_callback>;
-
-using websocket_close_code = boost::beast::websocket::close_code;
-
-http_server_ptr<http_stream_type> make_http_server(service_ptr as,
-                                                   string_view addr,
-                                                   uint16_t port = 80,
-                                                   bool reuse_port = false);
-http_server_ptr<https_stream_type> make_https_server(service_ptr,
-                                                     ssl::context&& ssl,
-                                                     string_view,
-                                                     uint16_t port = 443,
-                                                     bool reuse_port = false);
-// http_connection_ptr make_http_connection(service_ptr as, string_view addr,
-// string_view port);
-websocket_ptr make_websocket_connection(service_ptr as, string_view url,
-                                        int timeout = 10);
-
-http_request_ptr http_easy_request(service_ptr as, string_view method,
-                                   string_view url);
-template <typename D>
-http_request_ptr http_easy_request(service_ptr as, string_view method,
-                                   string_view url, D&& data);
-
-template <typename D>
-http_request_ptr http_easy_request(
-    service_ptr as, string_view method, string_view url, D&& data,
-    const std::map<string_view, string_view>& headers);
-
-template <typename D>
-http_request_ptr http_easy_request(
-    service_ptr as, int timeout_ms, string_view method, string_view url,
-    D&& data, const std::map<string_view, string_view>& headers);
-
-struct http_url_scheme {
-  boost::urls::url_view get_url_view() const { return uv; }
-
-  bool is_ssl() const { return is_ssl_; }
-  string_view host() const
-  {
-    return string_view(uv.encoded_host().data(), uv.encoded_host().length());
-  }
-  string_view authority() const
-  {
-    return string_view(uv.authority().data(), uv.authority().size());
-  }
-  string_view username() const { return username_; }
-  string_view password() const { return password_; }
-  uint16_t port() const { return port_; }
-  string_view target() const
-  {
-    return string_view(uv.encoded_resource().data(),
-                       uv.encoded_resource().size());
-  }
-
- private:
-  boost::urls::url_view uv;
-  std::string username_;
-  std::string password_;
-  bool is_ssl_;
-  uint16_t port_;
-
-  friend bool http_analyze_url(string_view u, http_url_scheme& scheme);
-};
-
-bool http_analyze_url(string_view url, http_url_scheme& scheme);
-
-namespace internal {
-std::string route_spec_to_regex(string_view route_spc);
-}
-
-template <typename StreamType>
-class http_server
-    : public std::enable_shared_from_this<http_server<StreamType>> {
- private:
-  struct private_ {};
-
- public:
-  using stream_type = StreamType;
-  ~http_server(){};
-  http_server& operator=(const http_server&) = delete;
-  http_server() = delete;
-  http_server(const http_server&) = delete;
-  http_server(http_server&&) = default;
-  http_server& operator=(http_server&&) = default;
-
-  http_server(struct private_&&, service_ptr as, string_view addr,
-              uint16_t port);
-
-  template <typename T>
-  void on_http_request(string_view route_spec, T&& cb)
-  {
-    std::regex re(internal::route_spec_to_regex(route_spec));
-    on_http_request_regex(re, "", std::forward<T>(cb));
-  }
-
-  template <typename T>
-  void on_http_request(string_view route_spec, string_view method, T&& cb)
-  {
-    std::regex re(internal::route_spec_to_regex(route_spec));
-    on_http_request_regex(re, method, std::forward<T>(cb));
-  }
-
-  template <typename R, typename M, typename T>
-  void on_http_request_regex(R&& r, M&& m, T&& cb)
-  {
-    http_routes.push_back({std::string{std::forward<M>(m)}, std::forward<R>(r),
-                           std::forward<T>(cb)});
-  }
-
-  template <typename T>
-  void on_websocket(string_view route_spec, T&& cb)
-  {
-    std::regex re(internal::route_spec_to_regex(route_spec));
-    on_websocket_regex(re, std::forward<T>(cb));
-  }
-
-  template <typename R, typename T>
-  void on_websocket_regex(R&& r, T&& cb)
-  {
-    ws_routes.push_back({"", std::forward<R>(r), std::forward<T>(cb)});
-  }
-
-  template <typename Req>
-  http_connection_ptr<stream_type> get_request_connection(const Req& req)
-  {
-    auto p =
-        boost::any_cast<http_connection_wptr<stream_type>>(req.connection_wptr);
-    if (auto connection = p.lock()) {
-      return connection;
-    }
-    return nullptr;
-  }
-
-  size_t get_request_header_limit() const { return request_header_limit; }
-
-  size_t get_request_body_limit() const { return request_body_limit; }
-
-  void set_request_header_limit(size_t l) { request_header_limit = l; }
-
-  void set_request_body_limit(size_t l) { request_body_limit = l; }
-  void close() { acceptor->close(); }
-
- private:
-  void start_accept(asio::io_context& io_service);
-
-  template <typename RouteType, typename ReqType>
-  const RouteType& find_route(const std::vector<RouteType>& routeList,
-                              const ReqType& req, http_route_args& a) const
-  {
-    auto it = find_if(routeList.begin(), routeList.end(),
-                      [&req, &a](const RouteType& tuple) -> bool {
-                        std::smatch m;
-                        auto method = std::get<0>(tuple);
-                        if (!method.compare("") ||
-                            boost::iequals(method, req.method_string())) {
-                          std::string s{req.target()};
-                          if (std::regex_search(s, m, std::get<1>(tuple))) {
-                            a.clear();
-                            for (const auto& item : m) a.push_back(item.str());
-                            return true;
-                          } else
-                            return false;
-                        } else
-                          return false;
-                      });
-    if (it != routeList.end())
-      return *it;
-    else
-      throw not_found_error("route not found");
-  }
-
-  template <typename ReqType>
-  const websocket_route_tuple& find_websocket_route(const ReqType& req,
-                                                    http_route_args& a) const
-  {
-    return find_route(ws_routes, req, a);
-  }
-
-  template <typename ReqType>
-  const http_route_tuple& find_http_route(const ReqType& req,
-                                          http_route_args& a) const
-  {
-    return find_route(http_routes, req, a);
-  }
-
-  std::shared_ptr<ip::tcp::acceptor> acceptor;
-  service_wptr service;
-
-  std::vector<http_route_tuple> http_routes;
-  std::vector<websocket_route_tuple> ws_routes;
-
-  std::shared_ptr<ssl::context> ssl_context;
-
-  size_t request_body_limit;
-  size_t request_header_limit;
-
-  template <typename S>
-  friend class http_connection;
-  friend http_server_ptr<http_stream_type> make_http_server(service_ptr,
-                                                            string_view,
-                                                            uint16_t, bool);
-  friend http_server_ptr<https_stream_type> make_https_server(
-      service_ptr, ssl::context&& ssl, string_view, uint16_t, bool);
-};
-
-template <typename StreamType>
-class http_connection
-    : public std::enable_shared_from_this<http_connection<StreamType>> {
- private:
-  struct private_ {};
-
- public:
-  ~http_connection(){};
-  http_connection& operator=(const http_connection&) = delete;
-  http_connection() = delete;
-  http_connection(const http_connection&) = delete;
-  http_connection(http_connection&&) = default;
-  http_connection& operator=(http_connection&&) = default;
-
-  // constructor for server connection
-  http_connection(struct private_&&, tcp::socket&& sock,
-                  http_server_ptr<http_stream_type> server)
-      : http_server(http_server_wptr<StreamType>(server)),
-        stream(std::move(sock)),
-        is_websocket(false),
-        is_server_connection(true),
-        remote_endpoint(
-            beast::get_lowest_layer(stream).socket().remote_endpoint()){};
-
-  http_connection(struct private_&&, tcp::socket&& sock,
-                  http_server_ptr<https_stream_type> server)
-      : http_server(http_server_wptr<StreamType>(server)),
-        ssl_context(server->ssl_context),
-        stream(std::move(sock), *ssl_context),
-        is_websocket(false),
-        is_server_connection(true),
-        remote_endpoint(
-            beast::get_lowest_layer(stream).socket().remote_endpoint()){};
-
-  // constructor for client connection
-  // template <typename executor_type>
-  // http_connection(struct private_ &&, const executor_type &io_service)
-  //     : http_server(nullptr),
-  //       socket(io_service),
-  //       stream(socket),
-  //       is_websocket(false),
-  //       is_server_connection(false){};
-
-  StreamType& get_stream() { return stream; };
-  tcp::endpoint get_remote_endpoint() const { return remote_endpoint; };
-
- private:
-  void start();
-  inline void handshake_if_ssl();
-  inline void shutdown_ssl();
-
-  http_server_wptr<StreamType> http_server;
-  std::shared_ptr<ssl::context> ssl_context;
-  StreamType stream;
-  bool is_websocket;
-  bool is_server_connection;
-  tcp::endpoint remote_endpoint;
-
-  template <typename S>
-  friend class http_server;
-  // friend http_connection_ptr make_http_connection(service_ptr as, string_view
-  // addr, string_view port);
-};
-
-class http_request : public std::enable_shared_from_this<http_request> {
- private:
-  struct private_ {};
-
- public:
-  ~http_request(){};
-  http_request& operator=(const http_request&) = delete;
-  http_request(const http_request&) = delete;
-  http_request(http_request&&) = default;
-  http_request& operator=(http_request&&) = default;
-
-  http_request()
-      : beast_request(),
-        headers(beast_request.base()),
-        body(beast_request.body()),
-        response(),
-        multipart_response(),
-        manual_response(false){};
-
-  http_beast_request beast_request;
-  http_request_headers& headers;
-  http_request_body& body;
-  boost::urls::url_view uv;
-
-  template <typename S>
-  inline auto get_connection_handle(S server)
-      -> decltype(server->get_request_connection(*this))
-  {
-    return server->get_request_connection(*this);
-  }
-
-  string_view target() const { return beast_request.target(); };
-
-  void target(string_view t) { beast_request.target(t); };
-
-  string_view method() const { return beast_request.method_string(); };
-
-  void method(string_view verb)
-  {
-    beast_request.method(beast::http::string_to_verb(verb));
-  };
-
-  void set_url_view() { uv = boost::urls::url_view(target()); }
-
-  boost::urls::url_view& get_url_view() { return uv; }
-
-  struct response {
-    response()
-        : beast_response(),
-          headers(beast_response.base()),
-          body(beast_response.body()){};
-    http_beast_response beast_response;
-    http_response_headers& headers;
-    http_response_body& body;
-    http_result result() const { return beast_response.result_int(); };
-    void result(http_result res) { beast_response.result(res); };
-  } response, multipart_response;
-
-  void activate_direct_response_handling() { manual_response = true; }
-
- private:
-  boost::beast::flat_buffer buffer;
-  bool manual_response;
-  boost::any connection_wptr;
-
-  template <typename StreamType>
-  friend class http_server;
-
-  template <typename StreamType>
-  friend class http_connection;
-
-  template <typename D, typename F>
-  friend http_request_ptr http_easy_request_multipart(
-      service_ptr as, int timeout_ms, string_view method, string_view url,
-      D&& data, const std::map<string_view, string_view>& headers, F&& f);
-
-  template <typename S, typename R, typename F>
-  friend boost::fibers::future<size_t> handle_client_request_response(
-      S& stream, int timeout_ms, R& req, http_url_scheme& scheme, F& f);
-};
-
-class websocket : public std::enable_shared_from_this<websocket> {
- public:
-  virtual ~websocket(){};
-  websocket& operator=(const websocket&) = delete;
-  websocket() = delete;
-  websocket(const websocket&) = delete;
-  websocket(websocket&&) = default;
-  websocket& operator=(websocket&&) = default;
-
- protected:
-  websocket(string_view host_, string_view port_, string_view path_)
-      : host(host_), port(port_), path(path_), is_server_connection(false){};
-
-  template <typename req_type>
-  websocket(req_type&& req)
-      : request(std::forward<req_type>(req)), is_server_connection(true){};
-
-  // API
- public:
-  virtual std::string get_string() = 0;
-  virtual size_t read_basic_buffer(std::vector<uint8_t>& b) = 0;
-  virtual void send_string(string_view s) = 0;
-  virtual void write_basic_buffer(const std::vector<uint8_t>& b) = 0;
-  virtual void set_keepalive_pings(bool b) = 0;
-  virtual void set_idle_timeout(int t) = 0;
-
-  void close(websocket_close_code code) { close(code, "NORMAL"); }
-  virtual void close(websocket_close_code code, string_view reason){};
-  http_request_ptr request;
-
- protected:
-  std::string host;
-  std::string port;
-  std::string path;
-
-  bool is_server_connection;
-};
-
-template <typename StreamType>
-class websocket_impl : public websocket {
- private:
-  struct private_ {};
-
- public:
-  virtual ~websocket_impl(){};
-  websocket_impl& operator=(const websocket_impl&) = delete;
-  websocket_impl() = delete;
-  websocket_impl(const websocket_impl&) = delete;
-  websocket_impl(websocket_impl&&) = default;
-  websocket_impl& operator=(websocket_impl&&) = default;
-
-  websocket_impl(struct private_&&, string_view host_, string_view port_,
-                 string_view path_)
-      : websocket(host_, port_, path_){};
-
-  template <typename ws_type, typename req_type>
-  websocket_impl(struct private_&&, ws_type&& ws, req_type&& req)
-      : websocket(std::forward<req_type>(req)), ws(std::forward<ws_type>(ws)){};
-
-  // API
-  virtual std::string get_string()
-  {
-    std::string message;
-    auto buffer = asio::dynamic_buffer(message);
-
-    // we need to call async ops thru strand for this op
-    // to be MT-safe
-    asio::dispatch(
-        ws->get_executor(),
-        use_fiber_future([ws_ = ws, &buffer]() -> fibers::future<size_t> {
-          return internal::websocket::async_read(*ws_, buffer);
-        }))
-        .get()
-        .get();
-
-    if (ws->got_binary())
-      throw asyik::unexpected_error(
-          "unexpected binary message when get_string()");
-
-    return message;
-  }
-
-  virtual void send_string(string_view s)
-  {
-    ws->binary(false);
-    internal::websocket::async_write(*ws, asio::buffer(s.data(), s.length()))
-        .get();
-  }
-
-  // basic buffer API
-  virtual size_t read_basic_buffer(std::vector<uint8_t>& b)
-  {
-    // auto buffer = asio::dynamic_buffer(b);
-    asio::mutable_buffers_1 mb(b.data(), b.size());
-    auto buffer = beast::buffers_adaptor<asio::mutable_buffers_1>(mb);
-
-    // we need to call async ops thru strand for this op
-    // to be MT-safe
-    return asio::dispatch(ws->get_executor(),
-                          use_fiber_future(
-                              [ws_ = ws, &buffer]() -> fibers::future<size_t> {
-                                return internal::websocket::async_read(*ws_,
-                                                                       buffer);
-                              }))
-        .get()
-        .get();
-  }
-
-  virtual void write_basic_buffer(const std::vector<uint8_t>& b)
-  {
-    ws->binary(true);
-    internal::websocket::async_write(*ws, asio::buffer(b.data(), b.size()))
-        .get();
-  }
-
-  virtual void set_keepalive_pings(bool b)
-  {
-    beast::websocket::stream_base::timeout opt;
-
-    ws->get_option(opt);
-    opt.keep_alive_pings = b;
-    ws->set_option(opt);
-  };
-
-  virtual void set_idle_timeout(int t)
-  {
-    beast::websocket::stream_base::timeout opt;
-
-    ws->get_option(opt);
-    opt.idle_timeout = std::chrono::seconds(t);
-    ws->set_option(opt);
-  };
-
-  virtual void close(websocket_close_code code, string_view reason)
-  {
-    const boost::beast::websocket::close_reason cr(code, reason);
-
-    try {
-      auto f = asio::dispatch(
-                   ws->get_executor(),
-                   use_fiber_future([ws_ = ws, cr]() -> fibers::future<void> {
-                     beast::get_lowest_layer(*ws_).cancel();
-                     return internal::websocket::async_close(*ws_, cr);
-                   }))
-                   .get();
-      if (f.wait_for(std::chrono::seconds(2)) ==
-          fibers::future_status::timeout) {
-        return;
-      }
-
-      f.get();
-    } catch (...) {
-    }
-  }
-
- private:
-  template <typename S>
-  friend class http_connection;
-  friend websocket_ptr make_websocket_connection(service_ptr as,
-                                                 string_view url, int timeout);
-
-  std::shared_ptr<StreamType> ws;
-};
+// Remaining template function implementations that need to stay in main header
 
 inline boost::optional<std::string> deduce_boundary(
     beast::string_view content_type)
@@ -752,7 +225,7 @@ template <typename D, typename F>
 http_request_ptr http_easy_request_multipart(
     service_ptr as, int timeout_ms, string_view method, string_view url,
     D&& data, const std::map<string_view, string_view>& headers, F&& f,
-    const digest_authenticator* auth = nullptr)
+    const digest_authenticator* auth)
 {
   http_url_scheme scheme;
